엑셀파일 .zip 변환, 관리자 페이지 페이징 및 포인트 비율

+++ src/main/java/com/tpay/domains/order/domain/OrderEntity.java
@@ -127,10 +127,9 @@ public class OrderEntity extends BaseTimeEntity {
         return points;
     }
 
-    public long getPointsWithPercentage(Integer balancePercentage){
+    public long getPointsWithPercentage(double balancePercentage){
         double vat = Double.parseDouble(this.totalVat);
         return (long) Math.floor((vat * balancePercentage) / 100);
-
     }
 
     public OrderEntity setOrderNumber(String orderNumber) {
